+++ indico/modules/events/abstracts/operations.py
@@ -81,7 +81,7 @@ def add_abstract_files(abstract, files, log_action=True):
             msg = 'Added file to abstract {}'.format(abstract.verbose_title)
         else:
             msg = 'Added {} files to abstract {}'.format(num, abstract.verbose_title)
-        abstract.event.log(EventLogRealm.management, EventLogKind.positive, 'Abstracts', msg, session.user,
+        abstract.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Abstracts', msg, session.user,
                            data={'Files': ', '.join(f.filename for f in abstract.files)})
 
 
@@ -96,7 +96,7 @@ def delete_abstract_files(abstract, files):
         msg = 'Deleted file from abstract {}'.format(abstract.verbose_title)
     else:
         msg = 'Deleted {} files from abstract {}'.format(num, abstract.verbose_title)
-    abstract.event.log(EventLogRealm.management, EventLogKind.negative, 'Abstracts', msg, session.user,
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Abstracts', msg, session.user,
                        data={'Files': ', '.join(f.filename for f in files)})
 
 
@@ -118,7 +118,7 @@ def create_abstract(event, abstract_data, custom_fields_data=None, send_notifica
     if send_notifications:
         send_abstract_notifications(abstract)
     logger.info('Abstract %s created by %s', abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.positive, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Abstracts',
                        'Abstract {} created'.format(abstract.verbose_title), session.user)
     return abstract
 
@@ -170,7 +170,7 @@ def update_abstract(abstract, abstract_data, custom_fields_data=None):
             'type': field_impl.log_type,
             'convert': lambda change, field_impl=field_impl: map(field_impl.get_friendly_value, change)
         }
-    abstract.event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
                        'Abstract {} modified'.format(abstract.verbose_title), session.user,
                        data={'Changes': make_diff_log(changes, log_fields)})
 
@@ -185,7 +185,7 @@ def withdraw_abstract(abstract):
     db.session.flush()
     signals.event.abstract_state_changed.send(abstract)
     logger.info('Abstract %s withdrawn by %s', abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.negative, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Abstracts',
                        'Abstract {} withdrawn'.format(abstract.verbose_title), session.user)
 
 
@@ -198,7 +198,7 @@ def delete_abstract(abstract, delete_contrib=False):
     db.session.flush()
     signals.event.abstract_deleted.send(abstract)
     logger.info('Abstract %s deleted by %s', abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.negative, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Abstracts',
                        'Abstract {} deleted'.format(abstract.verbose_title), session.user)
 
 
@@ -238,7 +238,7 @@ def judge_abstract(abstract, abstract_data, judgment, judge, contrib_session=Non
     if send_notifications:
         log_data['Notifications sent'] = send_abstract_notifications(abstract)
     logger.info('Abstract %s judged by %s', abstract, judge)
-    abstract.event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
                        'Abstract {} judged'.format(abstract.verbose_title), judge, data=log_data)
 
 
@@ -285,7 +285,7 @@ def update_reviewed_for_tracks(abstract, tracks):
     _update_tracks(abstract, tracks, only_reviewed_for=True)
     db.session.flush()
     logger.info('Reviewed tracks of abstract %s updated by %s', abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
                        'Reviewed tracks of abstract {} updated'.format(abstract.verbose_title),
                        session.user)
 
@@ -294,7 +294,7 @@ def reset_abstract_state(abstract):
     abstract.reset_state()
     db.session.flush()
     logger.info('Abstract %s state reset by %s', abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
                        'State of abstract {} reset'.format(abstract.verbose_title),
                        session.user)
 
@@ -305,7 +305,7 @@ def create_abstract_comment(abstract, comment_data):
     comment.abstract = abstract
     db.session.flush()
     logger.info("Abstract %s received a comment from %s", abstract, session.user)
-    abstract.event.log(EventLogRealm.management, EventLogKind.positive, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Abstracts',
                        'Abstract {} received a comment'.format(abstract.verbose_title),
                        session.user)
 
@@ -316,7 +316,7 @@ def update_abstract_comment(comment, comment_data):
     comment.modified_dt = now_utc()
     db.session.flush()
     logger.info("Abstract comment %s modified by %s", comment, session.user)
-    comment.abstract.event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    comment.abstract.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
                                'Comment on abstract {} modified'.format(comment.abstract.verbose_title),
                                session.user,
                                data={'Changes': make_diff_log(changes, {'text': 'Text', 'visibility': 'Visibility'})})
@@ -326,7 +326,7 @@ def delete_abstract_comment(comment):
     comment.is_deleted = True
     db.session.flush()
     logger.info("Abstract comment %s deleted by %s", comment, session.user)
-    comment.abstract.event.log(EventLogRealm.management, EventLogKind.negative, 'Abstracts',
+    comment.abstract.event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Abstracts',
                                'Comment on abstract {} removed'.format(comment.abstract.verbose_title),
                                session.user)
 
@@ -353,7 +353,7 @@ def create_abstract_review(abstract, track, user, review_data, questions_data):
         log_data['Other tracks'] = sorted(t.title for t in review.proposed_tracks)
     elif review.proposed_action in {AbstractAction.mark_as_duplicate, AbstractAction.merge}:
         log_data['Other abstract'] = review.proposed_related_abstract.verbose_title
-    abstract.event.log(EventLogRealm.management, EventLogKind.positive, 'Abstracts',
+    abstract.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Abstracts',
                        'Abstract {} reviewed'.format(abstract.verbose_title), user, data=log_data)
     return review
 
@@ -400,7 +400,7 @@ def update_abstract_review(review, review_data, questions_data):
             'title': 'Other tracks',
             'convert': lambda change: [sorted(t.title for t in x) for x in change]
         }
-    event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts',
+    event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts',
               'Review for abstract {} modified'.format(review.abstract.verbose_title),
               session.user, data={'Track': review.track.title, 'Changes': make_diff_log(changes, log_fields)})
 
@@ -415,17 +415,17 @@ def schedule_cfa(event, start_dt, end_dt, modification_end_dt):
         log_data['End'] = end_dt.isoformat()
     if modification_end_dt:
         log_data['Modification deadline'] = modification_end_dt.isoformat()
-    event.log(EventLogRealm.management, EventLogKind.change, 'Abstracts', 'Call for abstracts scheduled', session.user,
+    event.log(EventLogRealm.reviewing, EventLogKind.change, 'Abstracts', 'Call for abstracts scheduled', session.user,
               data=log_data)
 
 
 def open_cfa(event):
     event.cfa.open()
     logger.info("Call for abstracts for %s opened by %s", event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.positive, 'Abstracts', 'Call for abstracts opened', session.user)
+    event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Abstracts', 'Call for abstracts opened', session.user)
 
 
 def close_cfa(event):
     event.cfa.close()
     logger.info("Call for abstracts for %s closed by %s", event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.negative, 'Abstracts', 'Call for abstracts closed', session.user)
+    event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Abstracts', 'Call for abstracts closed', session.user)