+++ indico/modules/events/papers/operations.py
@@ -52,7 +52,7 @@ def set_reviewing_state(event, reviewing_type, enable):
     event.cfp.set_reviewing_state(reviewing_type, enable)
     action = 'enabled' if enable else 'disabled'
     logger.info("Reviewing type '%s' for event %r %s by %r", reviewing_type.name, event, action, session.user)
-    event.log(EventLogRealm.management, EventLogKind.positive if enable else EventLogKind.negative, 'Papers',
+    event.log(EventLogRealm.reviewing, EventLogKind.positive if enable else EventLogKind.negative, 'Papers',
               "{} {} reviewing".format("Enabled" if enable else "Disabled", orig_string(reviewing_type.title.lower())),
               session.user)
 
@@ -112,7 +112,7 @@ def update_team_members(event, managers, judges, content_reviewers=None, layout_
 def create_competences(event, user, competences):
     PaperCompetence(event=event, user=user, competences=competences)
     logger.info("Competences for user %r for event %r created by %r", user, event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+    event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
               "Added competences of {}".format(user.full_name), session.user,
               data={'Competences': ', '.join(competences)})
 
@@ -121,7 +121,7 @@ def update_competences(user_competences, competences):
     event = user_competences.event
     user_competences.competences = competences
     logger.info("Competences for user %r in event %r updated by %r", user_competences.user, event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+    event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
               "Updated competences for user {}".format(user_competences.user.full_name), session.user,
               data={'Competences': ', '.join(competences)})
 
@@ -134,20 +134,20 @@ def schedule_cfp(event, start_dt, end_dt):
     if end_dt:
         log_data['End'] = end_dt.isoformat()
     logger.info("Call for papers for %r scheduled by %r", event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.change, 'Papers', 'Call for papers scheduled', session.user,
+    event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers', 'Call for papers scheduled', session.user,
               data=log_data)
 
 
 def open_cfp(event):
     event.cfp.open()
     logger.info("Call for papers for %r opened by %r", event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.positive, 'Papers', 'Call for papers opened', session.user)
+    event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers', 'Call for papers opened', session.user)
 
 
 def close_cfp(event):
     event.cfp.close()
     logger.info("Call for papers for %r closed by %r", event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.negative, 'Papers', 'Call for papers closed', session.user)
+    event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Papers', 'Call for papers closed', session.user)
 
 
 def create_paper_revision(paper, submitter, files):
@@ -162,7 +162,7 @@ def create_paper_revision(paper, submitter, files):
     db.session.expire(revision._contribution, ['_paper_last_revision'])
     notify_paper_revision_submission(revision)
     logger.info('Paper revision %r submitted by %r', revision, session.user)
-    paper.event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
                     "Paper revision {} submitted for contribution {} ({})"
                     .format(revision.id, paper.contribution.title, paper.contribution.friendly_id), session.user)
     return revision
@@ -183,7 +183,7 @@ def judge_paper(paper, judgment, comment, judge):
     log_data = {'New state': orig_string(judgment.title)}
     notify_paper_judgment(paper)
     logger.info('Paper %r was judged by %r to %s', paper, judge, orig_string(judgment.title))
-    paper.event.log(EventLogRealm.management, EventLogKind.change, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers',
                     'Paper "{}" was judged'.format(orig_string(paper.verbose_title)), judge,
                     data=log_data)
 
@@ -193,7 +193,7 @@ def reset_paper_state(paper):
     db.session.flush()
     notify_paper_judgment(paper, reset=True)
     logger.info('Paper %r state reset by %r', paper, session.user)
-    paper.event.log(EventLogRealm.management, EventLogKind.change, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers',
                     'Judgment {} reset'.format(paper.verbose_title), session.user)
 
 
@@ -256,10 +256,10 @@ def update_reviewing_roles(event, users, contributions, role, assign):
         for user in users:
             notify_paper_assignment(user, role, contributions, event, assign)
     if assign:
-        event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+        event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
                   'Papers assigned ({})'.format(orig_string(role.title)), session.user, data=log_data)
     else:
-        event.log(EventLogRealm.management, EventLogKind.negative, 'Papers',
+        event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Papers',
                   'Papers unassigned ({})'.format(orig_string(role.title)), session.user, data=log_data)
     db.session.flush()
     logger.info('Paper reviewing roles in event %r updated by %r', event, session.user)
@@ -281,7 +281,7 @@ def create_review(paper, review_type, user, review_data, questions_data):
         'Action': orig_string(review.proposed_action.title),
         'Comment': review.comment
     })
-    paper.event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
                     'Paper for contribution {} reviewed'.format(paper.contribution.verbose_title),
                     user, data=log_data)
     return review
@@ -313,7 +313,7 @@ def update_review(review, review_data, questions_data):
         'proposed_action': 'Action',
         'comment': 'Comment'
     })
-    event.log(EventLogRealm.management, EventLogKind.change, 'Papers',
+    event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers',
               'Review for paper {} modified'.format(paper.verbose_title),
               session.user, data={'Changes': make_diff_log(changes, log_fields)})
 
@@ -334,7 +334,7 @@ def create_comment(paper, text, visibility, user):
     for receiver in recipients:
         notify_comment(receiver, paper, text, user)
     logger.info("Paper %r received a comment from %r", paper, session.user)
-    paper.event.log(EventLogRealm.management, EventLogKind.positive, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.positive, 'Papers',
                     'Paper {} received a comment'.format(paper.verbose_title),
                     session.user)
 
@@ -346,7 +346,7 @@ def update_comment(comment, text, visibility):
     db.session.flush()
     logger.info("Paper comment %r modified by %r", comment, session.user)
     paper = comment.paper_revision.paper
-    paper.event.log(EventLogRealm.management, EventLogKind.change, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers',
                     'Comment on paper {} modified'.format(paper.verbose_title), session.user,
                     data={'Changes': make_diff_log(changes, {'text': 'Text', 'visibility': 'Visibility'})})
 
@@ -356,7 +356,7 @@ def delete_comment(comment):
     db.session.flush()
     logger.info("Paper comment %r deleted by %r", comment, session.user)
     paper = comment.paper_revision.paper
-    paper.event.log(EventLogRealm.management, EventLogKind.negative, 'Papers',
+    paper.event.log(EventLogRealm.reviewing, EventLogKind.negative, 'Papers',
                     'Comment on paper {} removed'.format(paper.verbose_title), session.user)
 
 
@@ -367,5 +367,5 @@ def set_deadline(event, role, deadline, enforce=True):
     })
     log_data = {'Enforced': enforce, 'Deadline': deadline.isoformat() if deadline else 'None'}
     logger.info('Paper reviewing deadline (%s) set in %r by %r', role.name, event, session.user)
-    event.log(EventLogRealm.management, EventLogKind.change, 'Papers',
+    event.log(EventLogRealm.reviewing, EventLogKind.change, 'Papers',
               "Paper reviewing deadline ({}) set".format(role.title), session.user, data=log_data)